refactor: split external provider tests into multiple files
